can able to remove image from database when editing

+++ Report-Generator-EntityFramework/Commands/DeleteReportCommand.cs
@@ -1,11 +1,11 @@
-﻿using Report_Generator_Domain.Commands;
+﻿using Microsoft.EntityFrameworkCore;
+using Report_Generator_Domain.Commands;
 using Report_Generator_EntityFramework.DTOs;
 
 namespace Report_Generator_EntityFramework.Commands
 {
     public class DeleteReportCommand : IDeleteReportCommand
     {
-
         private readonly ReportModelDbContextFactory _contextFactory;
 
         public DeleteReportCommand(ReportModelDbContextFactory contextFactory)
@@ -17,14 +17,22 @@ namespace Report_Generator_EntityFramework.Commands
         {
             using (ReportModelDbContext context = _contextFactory.Create())
             {
-                ReportModelDto reportModelDto = new ReportModelDto()
+                // Find the report model by its ID
+                ReportModelDto reportModelDto = await context.ReportModels
+                    .Include(report => report.Images) // Include related images
+                    .FirstOrDefaultAsync(report => report.Id == id);
+
+                if (reportModelDto != null)
                 {
-                    Id = id,
+                    // Remove all associated images
+                    context.ReportImageModels.RemoveRange(reportModelDto.Images);
 
-                };
+                    // Remove the report model itself
+                    context.ReportModels.Remove(reportModelDto);
 
-                context.ReportModels.Remove(reportModelDto);
-                await context.SaveChangesAsync();
+                    // Save changes to the database
+                    await context.SaveChangesAsync();
+                }
             }
         }
     }
